test: delete empty enzyme tests; promote TL tests
